draytek_arsenal/requirements.txt
kaitaistruct==0.10
PyYAML==6.0.1
requests==2.31.0
